Add support for running coverage for integration tests

+++ Rakefile
@@ -10,6 +10,7 @@ OPENCOVER_CONSOLE = Dir['packages/OpenCover.*/OpenCover.Console.exe'].first
 REPORT_GENERATOR = Dir['packages/ReportGenerator.*/ReportGenerator.exe'].first
 
 UNIT_TESTS_DLL = "StyletUnitTests/bin/#{CONFIG}/StyletUnitTests.dll"
+INTEGRATION_TESTS_EXE = "StyletIntegrationTests/bin/#{CONFIG}/StyletIntegrationTests.exe"
 
 COVERAGE_DIR = 'Coverage'
 COVERAGE_FILE = File.join(COVERAGE_DIR, 'coverage.xml')
@@ -43,10 +44,49 @@ task :nunit do |t|
   sh NUNIT_EXE, UNIT_TESTS_DLL
 end
 
-desc "Generate code coverage reports for CONFIG (or Debug)"
-task :cover => [:build, COVERAGE_DIR] do |t|
-  sh OPENCOVER_CONSOLE, %Q{-register:user -target:"#{NUNIT_CONSOLE}" -filter:"+[Stylet]* -[Stylet]XamlGeneratedNamespace.*" -targetargs:"#{UNIT_TESTS_DLL} /noshadow" -output:"#{COVERAGE_FILE}"}
-  sh REPORT_GENERATOR, %Q{-reports:"#{COVERAGE_FILE}" "-targetdir:#{COVERAGE_DIR}"}
-  rm 'TestResult.xml'
+
+namespace :cover do
+
+  desc "Generate unit test code coverage reports for CONFIG (or Debug)"
+  task :unit => [:build, COVERAGE_DIR] do |t|
+    coverage(instrument(:nunit, UNIT_TESTS_DLL))
+  end
+
+  desc "Create integration test code coverage reports for CONFIG (or Debug)"
+  task :integration => [:build, COVERAGE_DIR] do |t|
+    coverage(instrument(:exe, INTEGRATION_TESTS_EXE))
+  end
+
+  desc "Create test code coverage for everything for CONFIG (or Debug)"
+  task :all => [:build, COVERAGE_DIR] do |t|
+    coverage([instrument(:nunit, UNIT_TESTS_DLL), instrument(:exe, INTEGRATION_TESTS_EXE)])
+  end
+
 end
+
+def instrument(runner, target)
+  case runner
+  when :nunit
+    opttarget = NUNIT_CONSOLE
+    opttargetargs = target
+  when :exe
+    opttarget = target
+    opttargetargs = ''
+  else
+    raise "Unknown runner #{runner}"
+  end
+ 
+  coverage_file = File.join(COVERAGE_DIR, File.basename(target).ext('xml'))
+  sh OPENCOVER_CONSOLE, %Q{-register:user -target:"#{opttarget}" -filter:"+[Stylet]* -[Stylet]XamlGeneratedNamespace.*" -targetargs:"#{opttargetargs} /noshadow" -output:"#{coverage_file}"}
+
+  rm 'TestResult.xml' if runner == :nunit
+
+  coverage_file
+end
+
+def coverage(coverage_files)
+  coverage_files = [*coverage_files]
+  sh REPORT_GENERATOR, %Q{-reports:"#{coverage_files.join(';')}" "-targetdir:#{COVERAGE_DIR}"}
+end
+
